Extract Verifier intergation test into abstract test
